[fix] 회원탈퇴시 users 컬렉션에서도 회원에 대한 도큐먼트를 삭제

+++ diEAT_SwiftUI/diEAT_SwiftUI/ViewModel/Authentication/AuthViewModel.swift
@@ -84,8 +84,12 @@ class AuthViewModel: ObservableObject {
     }
     
     func deleteUser() {
-        Auth.auth().currentUser?.delete() { _ in
-            self.userSession = nil
+        guard let uid = userSession?.uid else { return }
+        
+        Firestore.firestore().collection("users").document(uid).delete() { _ in
+            Auth.auth().currentUser?.delete() { _ in
+                self.logout()
+            }
         }
     }
